+++ spec/fixtures/SpectaFocus/FocusedSpec.m
@@ -1,0 +1,16 @@
+#import "Specta.h"
+
+SpecBegin(Focus)
+
+fcontext(@"focused context", ^{
+  fdescribe(@"focused describe", ^{
+    fexample(@"focused example" ^{
+    });
+    fit(@"focused it", ^{
+    });
+    fspecify(@"focused specify", ^{
+    });
+  });
+});
+
+SpecEnd