+++ spec/fixtures/SpectaFocus/InconsistentlyFormattedSpec.m
@@ -1,0 +1,10 @@
+#import "Specta.h"
+
+SpecBegin(Focus)
+  fdescribe (@"focused describe with space before the opening parenthesis" ^{
+    fit  (  @"focused it with two spaces before and after the opening parenthesis" ^{
+    })
+  })
+});
+
+SpecEnd