+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [#10](https://github.com/dblock/obcd/pull/10): Change output format to match what Xcode expects in order to display warnings - [@interstateone](https://github.com/interstateone).
 * [#11](https://github.com/dblock/obcd/pull/11): Fix: incorrect violation count with multiple violations from a single check - [@interstateone](https://github.com/interstateone).
 * [#11](https://github.com/dblock/obcd/pull/11): Fix: HeaderStyle incorrectly detects blank line after copyright - [@interstateone](https://github.com/interstateone).
+* [#15](https://github.com/dblock/obcd/pull/15): Added [SpectaFocus](docs/checks/SpectaFocus.md) check - [@w5mith](https://github.com/w5mith).
 
 ### 0.1.0 (3/28/2014)